+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - `editor`
   - [#735](https://github.com/wix-incubator/rich-content/pull/735) disable paste inline images and atomic blocks
   - [#735](https://github.com/wix-incubator/rich-content/pull/735) copy paste inline styles from html
-  
+  - [#713](https://github.com/wix-incubator/rich-content/pull/713) toolbars rtl issued fixed
 </details>
 <hr/>